refactor(frontend): migrate watch tests to useresourcewatch tests
Migrate the tests for Watch to the useResourceWatch tests
