+++ frontend/src/methods/useResourceWatch.spec.ts
@@ -2,13 +2,21 @@
 //
 // Use of this software is governed by the Business Source License
 // included in the LICENSE file.
-import { createBootstrapEvent, createCreatedEvent } from '@msw/helpers'
+import {
+  createBootstrapEvent,
+  createCreatedEvent,
+  createDestroyedEvent,
+  createUpdatedEvent,
+} from '@msw/helpers'
 import { createWatchStreamMock } from '@msw/server'
 import { waitFor } from '@testing-library/vue'
 import { describe, expect, test } from 'vitest'
 import { effectScope, nextTick, ref } from 'vue'
 
 import { Runtime } from '@/api/common/omni.pb'
+import type { Resource } from '@/api/grpc'
+import type { MachineSpec } from '@/api/omni/specs/omni.pb'
+import { DefaultNamespace, MachineType } from '@/api/resources'
 
 import { useResourceWatch } from './useResourceWatch'
 
@@ -23,6 +31,103 @@ function renderComposable<T>(factory: () => T) {
 }
 
 describe('useResourceWatch', () => {
+  test('handles created, updated, and destroyed events across bootstrap', async () => {
+    const { pushEvents } = createWatchStreamMock({ skipBootstrap: true })
+
+    const { data, loading, unmount } = renderComposable(() =>
+      useResourceWatch<MachineSpec>({
+        runtime: Runtime.Omni,
+        resource: { type: MachineType, namespace: DefaultNamespace },
+      }),
+    )
+
+    const machine1: Resource<MachineSpec> = {
+      metadata: { id: '1', namespace: 'default', type: MachineType },
+      spec: { connected: true, management_address: 'localhost' },
+    }
+
+    const machine2: Resource<MachineSpec> = {
+      metadata: { id: '2', namespace: 'default', type: MachineType },
+      spec: { connected: true, management_address: 'localhost' },
+    }
+
+    await pushEvents(
+      createCreatedEvent(machine1),
+      createCreatedEvent(machine2),
+      createDestroyedEvent(machine2),
+      createUpdatedEvent(
+        {
+          ...machine1,
+          spec: { ...machine1.spec, connected: false },
+        },
+        machine1,
+      ),
+    )
+
+    // not yet bootstrapped
+    expect(data.value).toHaveLength(0)
+
+    // still loading
+    expect(loading.value).toBeTruthy()
+
+    // Bootstrap event triggers the loading of queued events
+    await pushEvents(createBootstrapEvent())
+
+    await waitFor(() => expect(data.value).toHaveLength(1))
+
+    const machine = data.value[0]
+    expect(machine.metadata.id).toBe('1')
+    expect(machine.spec.connected).toBe(false)
+    expect(loading.value).toBe(false)
+
+    await pushEvents(createCreatedEvent(machine2))
+
+    await waitFor(() => expect(data.value).toHaveLength(2))
+
+    unmount()
+
+    expect(data.value).toHaveLength(0)
+  })
+
+  test('resets and reloads items after stream reconnect', async () => {
+    const { pushEvents, closeStream } = createWatchStreamMock({ skipBootstrap: true })
+
+    const { data, loading } = renderComposable(() =>
+      useResourceWatch<MachineSpec>({
+        runtime: Runtime.Omni,
+        resource: { type: MachineType, namespace: DefaultNamespace },
+      }),
+    )
+
+    async function populate(count: number) {
+      await pushEvents(
+        ...new Array(count).fill(null).map((_, i) =>
+          createCreatedEvent<MachineSpec>({
+            metadata: { id: i.toString(), namespace: 'default', type: MachineType },
+            spec: { connected: true, management_address: 'localhost' },
+          }),
+        ),
+        createBootstrapEvent(),
+      )
+    }
+
+    await populate(4)
+
+    await waitFor(() => {
+      expect(loading.value).toBe(false)
+      expect(data.value).toHaveLength(4)
+    })
+
+    await closeStream(new Error('network error'))
+
+    await populate(2)
+
+    await waitFor(() => {
+      expect(loading.value).toBe(false)
+      expect(data.value).toHaveLength(2)
+    })
+  })
+
   test('updates single resource data after watch events', async () => {
     const { pushEvents } = createWatchStreamMock({ skipBootstrap: true })
